Working on Framebase converter
