+++ src/main/config/premonitor.properties
@@ -138,3 +138,15 @@ resource11.linkfn = fn16
 resource11.language = en
 resource11.web = https://github.com/newsreader/eso/
 resource11.active = 1
+
+resource12.label = FrameBase 2.0
+resource12.class = eu.fbk.dkm.premon.premonitor.FramebaseConverter
+resource12.resource = fb
+resource12.source = fb20
+resource12.folder = resources/framebase-2.0
+resource12.linkfn = fn16
+resource12.linkpb = pb17,pb215
+resource12.linknb = nb10
+resource12.language = en
+resource12.web = http://www.framebase.org/
+resource12.active = 1
